Refined classes plus tests

+++ Model/Quizes/Poll.php
@@ -2,6 +2,7 @@
 /* vim: set expandtab tabstop=4 shiftwidth=4: */
 namespace Egulias\QuizBundle\Model\Quizes;
 
+use Egulias\QuizBundle\Model\Questions\Question;
 /**
  * Declaration of Poll quiz type
  * @author Eduardo Gulias Davis <me@egulias.com>
@@ -21,12 +22,28 @@ class Poll extends Quiz
 
     public function addQuestion(Question $q)
     {
-        $this->questions->detach($this->questions->current());
-        return parent::addQuestion($q);
+        parent::addQuestion($q);
+        if($this->questions->count() > 1)
+        {
+            $this->questions->rewind();
+            $this->questions->detach($this->questions->current());
+        }
+        return $this;
     }
 
     public function check()
     {
         return $this->questions->current()->check();
+    }
+
+    /**
+     *
+     *
+     * @see Quiz
+     */
+    public function getQuestion()
+    {
+        $this->questions->rewind();
+        return $this->questions->current();
     }
 }
